+++ CHANGES.md
@@ -1,3 +1,5 @@
+* Issue #965: output genome sizes with and without padding into Log.out.
+* Issue #955: fixed seg-fault-causing bug for --soloFeatures SJ option.
 * Docker build: switched to debian:stable-slim in the Dockerfile.
 * --soloType CB_samTagOut now allows output of (uncorrected) UMI sequences and quality scores with SAM tags UR and UY.